Make Histogram::record_many O(1) end to end
record_many(value, count) is documented as recording a value multiple
times, but every implementation degraded to the default trait impl: a
loop of count individual record() calls. In metrics-exporter-prometheus
each of those calls buffers a (value, Instant) sample until the next
upkeep, so record_many(v, 10_000_000) allocated tens of millions of
buffered samples. This made record_many unusable as a weighted observe
(e.g. time-weighted occupation histograms recording microseconds spent
at a level).

Changes, per crate:

- metrics: HistogramFn for Arc<T> forwards record_many to the inner
  impl so wrapper chains cannot silently degrade to the loop.
- metrics-util: Generational<T> and FanoutHistogram forward
  record_many; storage::Histogram grows record_n (O(buckets));
  storage::Summary grows add_n backed by DDSketch::add_with_count
  (sketches-ddsketch 0.4).
- metrics-exporter-prometheus: the registry buffers one
  (value, count, timestamp) entry per record_many call, and all three
  distribution types apply counts in constant time on drain: bucketed
  histograms via record_n, summaries via RollingSummary::add_n ->
  Summary::add_n, native histograms via observe_n.

Every path is O(1) in count, including the default summary
distribution, so no configuration can silently reintroduce the loop.

Blocked on a sketches-ddsketch release containing add_with_count.

+++ metrics-exporter-prometheus/src/protobuf.rs
@@ -158,7 +158,7 @@ pub(crate) fn render_protobuf_to_write<W: Write>(
                     pb::Metric {
                         label: label_pairs,
                         summary: Some(pb::Summary {
-                            sample_count: Some(summary.count() as u64),
+                            sample_count: Some(summary.count()),
                             sample_sum: Some(sum),
                             quantile: quantile_values,
